Noise function now working
After testing against a copy of the `noise.m` program running in Matlab
I am confident that the noise function now works the same way in Julia.
This is a small but positive step toward a full port as this was one of
the most difficult parts of the program to translate accurately.